Added python script to check various R1, R2, R3 proportions
Created a python script to run through the hysteresis transistion points
for various proportions of R1, R2 and R3. This helped in finding the
best proportions of R1 to R2 to R3 for the largest hysteresis transition voltages.

Showed a misinterpertation in the notes previously commited which were corrected in this commit as well.

An LTSPICE simulation of positive feedback hysteresis resistors was also
added to verify notes on this sections of the circuit.